pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "gps-tracker"
version = "0.1.0"
description = "GPS tracking application with PyQt6 GUI, serial GPS and IP geolocation support"
requires-python = ">=3.9"
dependencies = [
    "PyQt6>=6.4.0",
    "pyserial>=3.5",
    "pynmea2>=1.18.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pylint>=2.17.0",
    "black>=23.3.0",
    "pytest>=7.3.1",
]

[project.scripts]
gps-tracker = "main:main"
